+++ src/main/typescript/data/item/weapon/ranges/source.ts
@@ -3,6 +3,9 @@ import { SpecialName, SpecialNames } from "../../../../constants.js";
 
 /** An interface that represents the source ranges of a weapon. */
 export default interface Ranges {
+  /** Whether the weapon uses Point Blank rules */
+  pointBlank?: boolean;
+
   /** The short range of the weapon */
   short: Range;
 
@@ -28,14 +31,10 @@ export interface Range {
   modifier: number;
 }
 
-/**
- * A distance specifier for a weapon range.
- */
+/** A distance specifier for a weapon range. */
 export type Distance = number | "melee" | SpecialBasedRange;
 
-/**
- * A SPECIAL based range
- */
+/** A SPECIAL based range */
 export interface SpecialBasedRange {
   /** The base range of the range in meters */
   base: number;
@@ -47,7 +46,6 @@ export interface SpecialBasedRange {
   special: SpecialName;
 }
 
-/** A default value for a SPECIAL based distance */
 /** A JSON schema for a single range object */
 const RANGE_JSON_SCHEMA: JSONSchemaType<Range> = {
   description: "A range definition",
@@ -121,6 +119,14 @@ export const RANGES_JSON_SCHEMA: JSONSchemaType<Ranges> = {
   description: "Ranges definitions",
   type: "object",
   properties: {
+    pointBlank: {
+      type: "boolean",
+      description:
+        "Whether this weapon is eligible for the Point Blank range rule. If " +
+        "this is not specified, it defaults to `false`.",
+      nullable: true,
+      default: true
+    },
     short: {
       ...RANGE_JSON_SCHEMA,
       description: "The short range specification"